Use momentary layer switching
Instead of having to do an entire 3 key presses to fire keys on
the special layer (one press to open the layer, one to fire the key
and one to return to the base layer), momentary layer switching means
that the key can be fired in a single chord.

This also makes it much easier to determine which layer you are on,
as if you are not holding any keys down, then you are on the base
layer. I developed a habit of pressing the "return to base layer"
key pattern quite a lot and I believe it was not good for my hands to
be typing without any need for it.